upgrading sdl2 bindings

+++ core/lib/openssl/openssl.cpp
@@ -336,7 +336,7 @@ extern "C" {
     bio = BIO_push(b64, bio);
 
     BIO_set_flags(bio, BIO_FLAGS_BASE64_NO_NL); //Do not use newlines to flush buffer
-    size_t length = BIO_read(bio, buffer, decodeLen);
+    int length = (int)BIO_read(bio, buffer, decodeLen);
     BIO_free_all(bio);
 
     APITools_SetStringValue(context, 0, BytesToUnicode(buffer));
